Move server time state into the server reducer
