+++ src/components/HeaderBar/Progress.jsx
@@ -2,7 +2,7 @@ import cx from 'clsx';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useStore } from '../../hooks/useRedux';
-import { currentTimeSelector } from '../../reducers/time';
+import { currentTimeSelector } from '../../reducers/server';
 
 const { useEffect, useRef } = React;
