+++ src/reducers/index.ts
@@ -12,7 +12,6 @@ import playlists from './playlists';
 import server from './server';
 import settings from './settings';
 import theme from './theme';
-import time from './time';
 import users from './users';
 import waitlist from './waitlist';
 import mobile from '../mobile/reducers';
@@ -32,7 +31,6 @@ export {
   server,
   settings,
   theme,
-  time,
   users,
   waitlist,
   mobile,
